Add every page to the sidebar
except for the glossary.

I found navigating two different menus (bottom and side) always confusing so I attempted to categorize everything in to a nested sidebar. Maybe these can be made collapsible in the future.

I did not change paths or filenames in order to preserve links even though some of these should probably be adjusted.